+++ sigle/src/components/Form.ts
@@ -6,11 +6,11 @@ export const FormRow = styled.div`
 `;
 
 export const FormLabel = styled.label`
-  ${tw`w-full block tracking-wide font-bold text-black mb-2`};
+  ${tw`w-full block tracking-wide font-bold mb-2`};
 `;
 
 export const FormInput = styled.input`
-  ${tw`block w-full bg-white border border-grey rounded py-3 px-3 text-sm leading-tight focus:border-grey-dark focus:ring-grey-dark`};
+  ${tw`block w-full bg-white dark:text-black border border-grey rounded py-3 px-3 text-sm leading-tight focus:border-grey-dark focus:ring-grey-dark`};
 `;
 
 export const FormInputCheckbox = styled.input`
@@ -18,7 +18,7 @@ export const FormInputCheckbox = styled.input`
 `;
 
 export const FormTextarea = styled.textarea`
-  ${tw`block w-full bg-white border border-grey rounded py-3 px-3 text-sm leading-tight focus:border-grey-dark focus:ring-grey-dark`};
+  ${tw`block w-full bg-white dark:text-black border border-grey rounded py-3 px-3 text-sm leading-tight focus:border-grey-dark focus:ring-grey-dark`};
 `;
 
 export const FormHelper = styled.p`