+++ sigle/src/ui/index.ts
@@ -9,3 +9,4 @@ export * from './IconButton';
 export * from './ScrollArea';
 export * from './Text';
 export * from './Tooltip';
+export * from './Tabs';