+++ sigle/src/utils/migrations/story.ts
@@ -1,10 +1,10 @@
-import { convertSlateToHTML } from '../../modules/editor/utils/convertStateToHTML';
+import { generateHtml } from '@sigle/slate-to-html';
 import { Story } from '../../types';
 
 export const migrationStory = (file: Story): Story => {
   // When contentVersion is not set, it means V1
   if (!file.contentVersion) {
-    file.content = convertSlateToHTML(file.content);
+    file.content = generateHtml(file.content);
     file.contentVersion = '2';
   }
   return file;